Add essential stages
