+++ src/main/java/com/github/srilakshmikanthanp/facsimile/Facsimile.java
@@ -11,8 +11,14 @@ import javafx.stage.*;
 import javafx.scene.layout.*;
 import javafx.scene.control.*;
 
-import com.github.srilakshmikanthanp.facsimile.panes.*;
+import com.github.srilakshmikanthanp.facsimile.datum.*;
+import com.github.srilakshmikanthanp.facsimile.stages.*;
+import com.github.srilakshmikanthanp.facsimile.system.*;
+import com.github.srilakshmikanthanp.facsimile.utility.*;
 
+/**
+ * Main Application class
+ */
 public class Facsimile extends Application 
 {
     @Override
