--- src/main/java/com/github/srilakshmikanthanp/facsimile/panes/PrefsPane.java
@@ -1,6 +1,0 @@
-package com.github.srilakshmikanthanp.facsimile.panes;
-
-public class PrefsPane 
-{
-    
-}
